feat: address review feedback for transformation engine foundation
- Move WasmRuntime to weavster-core to enable functional unit testing
- Replace mock test implementations with actual WASM execution
- Upgrade record counts to i64/BIGINT to prevent integer overflow
- Improve Jinja substitution with robust regex-based replacement
- Fix non-deterministic and unsafe code generation in add_fields transform
- Introduce ProcessingStatus enum for strongly-typed state management
- Centralize SQLite URL construction and deduplicate error handling
- Clean up redundant files (GEMINI.md, check_eq.rs) and update .gitignore

+++ crates/weavster-runtime/migrations_sqlite/20260310000000_init.sql
@@ -4,7 +4,7 @@ CREATE TABLE processed_files (
     file_path TEXT NOT NULL,
     file_hash TEXT NOT NULL,
     processed_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    record_count INTEGER NOT NULL,
+    record_count BIGINT NOT NULL,
     status TEXT NOT NULL,
     error_message TEXT,
     UNIQUE(flow_name, file_path, file_hash)
@@ -28,8 +28,8 @@ CREATE TABLE flow_executions (
     started_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     completed_at TIMESTAMP,
     status TEXT NOT NULL,
-    records_processed INTEGER NOT NULL DEFAULT 0,
-    records_failed INTEGER NOT NULL DEFAULT 0,
+    records_processed BIGINT NOT NULL DEFAULT 0,
+    records_failed BIGINT NOT NULL DEFAULT 0,
     error_message TEXT
 );
